[feature #161402923] ensured express was starting the server moved all routes to app.js
